core: sim-infra: move block recovery code to a separate file
