+++ core/kt-osrd-sim-infra/src/main/kotlin/fr/sncf/osrd/sim_infra/api/LoadedSignalingInfra.kt
@@ -64,124 +64,3 @@ interface BlockInfra {
     fun getBlocksAtSignal(signal: LogicalSignalId): StaticIdxList<Block>
     fun getSignalsPositions(block: BlockId): DistanceList
 }
-
-data class BlockPathElement(
-    val prev: BlockPathElement?,
-    val block: BlockId,
-    val route: RouteId,
-    val routeStartOffset: Int,
-    val routeEndOffset: Int,
-)
-
-fun BlockPathElement.toList() : StaticIdxList<Block> {
-    val res = mutableStaticIdxArrayListOf(this.block)
-    var cur = this.prev
-    while (cur != null) {
-        res.add(cur.block)
-        cur = cur.prev
-    }
-    return res.reversed()
-}
-
-
-fun filterBlocks(
-    allowedSignalingSystems: StaticIdxList<SignalingSystem>,
-    blockInfra: BlockInfra,
-    blocks: StaticIdxList<Block>,
-    routePath: StaticIdxList<ZonePath>,
-    routeOffset: Int,
-): StaticIdxList<Block> {
-    val remainingZonePaths = routePath.size - routeOffset
-    val res = mutableStaticIdxArrayListOf<Block>()
-    blockLoop@ for (block in blocks) {
-        if (!allowedSignalingSystems.contains(blockInfra.getBlockSignalingSystem(block)))
-            continue
-        val blockPath = blockInfra.getBlockPath(block)
-        if (blockPath.size > remainingZonePaths)
-            continue
-        for (i in 0 until blockPath.size)
-            if (routePath[routeOffset + i] != blockPath[i])
-                continue@blockLoop
-        res.add(block)
-    }
-    return res
-}
-
-
-fun findRouteBlocks(
-    signalingInfra: RawSignalingInfra,
-    blockInfra: BlockInfra,
-    allowedSignalingSystems: StaticIdxList<SignalingSystem>,
-    previousPaths: List<BlockPathElement>?,
-    route: RouteId,
-): List<BlockPathElement> {
-    val routePath = signalingInfra.getRoutePath(route)
-    var maxRouteEndOffset = 0
-    val incompletePaths = PriorityQueue<BlockPathElement>(Comparator.comparing { it.routeEndOffset })
-    val completePaths = mutableListOf<BlockPathElement>()
-
-    fun addPath(path: BlockPathElement) {
-        if (path.routeEndOffset == routePath.size) {
-            completePaths.add(path)
-            return
-        }
-        if (path.routeEndOffset > maxRouteEndOffset)
-            maxRouteEndOffset = path.routeEndOffset
-        incompletePaths.add(path)
-    }
-
-    fun findNextBlocks(prevPath: BlockPathElement, routeOffset: Int) {
-        val lastBlock = prevPath.block
-        if (blockInfra.blockStopAtBufferStop(lastBlock))
-            return
-        val blockSignals = blockInfra.getBlockSignals(lastBlock)
-        val curSignal = blockSignals[blockSignals.size - 1]
-        val blocks = blockInfra.getBlocksAtSignal(curSignal)
-        val blocksOnRoute = filterBlocks(allowedSignalingSystems, blockInfra, blocks, routePath, routeOffset)
-        for (block in blocksOnRoute) {
-            val blockSize = blockInfra.getBlockPath(block).size
-            addPath(BlockPathElement(prevPath, block, route, routeOffset, routeOffset + blockSize))
-        }
-    }
-
-    // initialize with the BlockPathElements which are acceptable at the start of the route
-    if (previousPaths == null) {
-        val currentDet = signalingInfra.getZonePathEntry(routePath[0])
-        val blocks = blockInfra.getBlocksAtDetector(currentDet)
-        val blocksOnRoute = filterBlocks(allowedSignalingSystems, blockInfra, blocks, routePath, 0)
-        for (block in blocksOnRoute) {
-            val blockPath = blockInfra.getBlockPath(block)
-            addPath(BlockPathElement(null, block, route, 0, blockPath.size))
-        }
-    } else {
-        for (prevPath in previousPaths)
-            findNextBlocks(prevPath, 0)
-    }
-
-    // for each block until the end of the route path,
-    // filter candidates which don't match and add new candidates
-    while (incompletePaths.isNotEmpty()) {
-        val candidatePath = incompletePaths.poll()!!
-        findNextBlocks(candidatePath, candidatePath.routeEndOffset)
-    }
-
-    assert(completePaths.isNotEmpty())
-
-    return completePaths
-}
-
-
-fun getRouteBlocks(
-    signalingInfra: RawSignalingInfra,
-    blockInfra: BlockInfra,
-    routes: StaticIdxList<Route>,
-    allowedSignalingSystems: StaticIdxList<SignalingSystem>
-): List<StaticIdxList<Block>> {
-    var candidateBlocks: List<BlockPathElement>? = null
-
-    for (route in routes) {
-        val newCandidateBlocks = findRouteBlocks(signalingInfra, blockInfra, allowedSignalingSystems, candidateBlocks, route)
-        candidateBlocks = newCandidateBlocks
-    }
-    return candidateBlocks!!.map { it.toList() }
-}
